P4PU-1098 renamed tags
