+++ src/main/java/it/gov/pagopa/pu/citizen/controller/InstallmentController.java
@@ -2,6 +2,7 @@ package it.gov.pagopa.pu.citizen.controller;
 
 import it.gov.pagopa.pu.citizen.controller.generated.InstallmentApi;
 import it.gov.pagopa.pu.citizen.dto.InstallmentDebtorExtendedDTO;
+import it.gov.pagopa.pu.citizen.dto.generated.DebtorUnpaidDebtPositionInstallmentsDTO;
 import it.gov.pagopa.pu.citizen.security.SecurityUtils;
 import it.gov.pagopa.pu.citizen.service.installment.InstallmentFacadeService;
 import lombok.extern.slf4j.Slf4j;
@@ -24,5 +25,11 @@ public class InstallmentController implements InstallmentApi {
   public ResponseEntity<List<InstallmentDebtorExtendedDTO>> getInstallmentsByIuvOrNav(Long brokerId, String iuvOrNav, String debtorFiscalCode, String orgFiscalCode) {
     log.info("Requested getInstallmentsByIuvOrNav having brokerId {} and iuvOrNav {} ", brokerId, iuvOrNav);
     return ResponseEntity.ok(installmentFacadeService.getInstallmentByIuvOrNav(brokerId,iuvOrNav,debtorFiscalCode,orgFiscalCode,SecurityUtils.getAccessToken()));
+  }
+
+  @Override
+  public ResponseEntity<List<DebtorUnpaidDebtPositionInstallmentsDTO>> getDebtorUnpaidDebtPositionInstallments(Long brokerId, Long debtPositionId, Long paymentOptionId, String xFiscalCode, Long organizationId) {
+    log.info("User requested getDebtorUnpaidDebtPositionInstallments having debtPositionId {} and paymentOptionId {} with  brokerId {} organizationId {}", debtPositionId, paymentOptionId, brokerId, organizationId);
+    return ResponseEntity.ok(installmentFacadeService.getDebtorInstallmentNoPII(brokerId, debtPositionId,paymentOptionId, xFiscalCode, organizationId, SecurityUtils.getAccessToken()));
   }
 }
