--- src/main/java/it/gov/pagopa/pu/citizen/service/installment/InstallmentRetrieverService.java
@@ -1,9 +1,0 @@
-package it.gov.pagopa.pu.citizen.service.installment;
-
-import it.gov.pagopa.pu.citizen.dto.generated.DebtorUnpaidDebtPositionInstallmentsDTO;
-
-import java.util.List;
-
-public interface InstallmentRetrieverService {
-  List<DebtorUnpaidDebtPositionInstallmentsDTO> getDebtorInstallmentNoPII(Long brokerId, Long debtPositionId, Long paymentOptionId, String xFiscalCode, Long organizationId, String accessToken);
-}
